bug#10033 Remove deprecations from scala.text
Move scala.text.Document into compiler module
(in scala.tools.nsc.ast.TreeBrowser) and un-deprecate.
